// generated/src/app/api-models/ipa-contracts.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class IpaContracts {
  seqIpaContract: number;
  ipaId: string;
  lineOfBusiness: string;
  effectiveDate: Date;
  termDate: Date;
  termReason: string;
  contractInformation: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  serviceAreaName: string;
}